requirements.txt
fastapi
gradio
uvicorn
langchain
langchain-community
langchain-openai
langchain-chroma
pandas
numpy
python-dotenv
